// cadriciel/client/src/app/enums/svg-attributes.ts
export enum SVGAttributes {
    CX = 'cx',
    CY = 'cy',
    Cursor = 'cursor',
    Fill = 'fill',
    FillOpacity = 'fill-opacity',
    Filter = 'filter',
    Font = 'font-family',
    FontStyle = 'font-style',
    FontWeight = 'font-weight',
    Size = 'font-size',
    Points = 'points',
    Outline = 'outline',
    R = 'r',
    RX = 'rx',
    RY = 'ry',
    DX = 'dx',
    DY = 'dy',
    TextAnchor = 'text-anchor',
    Stroke = 'stroke',
    StrokeAlignment = 'stroke-alignment',
    StrokeDasharray = 'stroke-dasharray',
    StrokeLineCap = 'stroke-linecap',
    StrokeLinejoin = 'stroke-linejoin',
    StrokeMiterLimit = 'stroke-miterlimit',
    StrokeOpacity = 'stroke-opacity',
    StrokeWidth = 'stroke-width',
    TextContent = 'textContent',
    Tspan = 'tspan',
    Style = 'style',
    Height = 'height',
    Width = 'width',
    X = 'x',
    Y = 'y',
    DPath = 'd',
    Transform = 'transform',
    Text = 'text',
    Rect = 'rect',
    SVG = 'svg',
    G = 'g',
    Polyline = 'polyline',
    Circle = 'circle',
    Ellipse = 'ellipse',
    Polygon = 'polygon',
    Path = 'path',
    Line = 'line',
    X1 = 'x1',
    X2 = 'x2',
    Y1 = 'y1',
    Y2 = 'y2',
    ShapeRendering = 'shape-rendering',
    CrispEdges = 'crispEdges',
    Display = 'display',
    None = 'none',
    Inline = 'inline',
    Opacity = 'opacity',
    Class = 'class',
    Pen = 'pen',
    Aerosol = 'aerosol',
    Feather = 'feather',
    Id = 'id',
    FillRule = 'fill-rule',
    D = 'd',
    EvenOdd = 'evenodd',
}